Enhance WRF backend configuration and model initialization
- Added a new model section in the forecast configuration file to specify time step and physics options.
- Updated the WRFModel class to initialize from the new configuration structure, improving model setup.
- Refactored CMakeLists.txt to define the WRF backend as an INTERFACE library and include necessary header files for IDE visibility.
- Removed the obsolete WRFModel.cpp file to streamline the backend implementation.
